split stardew rules in multiple files
