+++ worlds/stardew_valley/stardew_rule/player_context.py
@@ -1,0 +1,6 @@
+class PlayerContext:
+    pass
+
+
+def fork_per_player(cls):
+    pass
